chore(channels/discord): merge outbound-stream/chunk tests into outbound.test.ts
Leftover from the brief stream-buffer.ts / message-chunk.ts split that
was already reverted (commit c7361ac merged the source files back into
outbound.ts). The 3 test files all targeted the same outbound.ts surface.

Now: 1 test file per source file (matches dedupe.test.ts, channel-history.test.ts
convention). Three describe blocks inside.
