+++ src/channels/discord/outbound.test.ts
@@ -1,7 +1,65 @@
 // Tests for the Discord outbound streaming pipeline.
 import { describe, it, expect, vi } from "vitest";
 import type { Message as DiscordMessage, SendableChannels } from "discord.js";
-import { createDiscordCallbacks } from "./outbound.js";
+import { SegmentedStreamBuffer, chunkDiscordMessage, createDiscordCallbacks } from "./outbound.js";
+
+describe("SegmentedStreamBuffer", () => {
+  it("does not flush below maxBufferSize threshold", async () => {
+    const onFlush = vi.fn().mockResolvedValue(undefined);
+    const buf = new SegmentedStreamBuffer(onFlush, 50);
+    await buf.append("Short text. With a boundary.");
+    expect(onFlush).not.toHaveBeenCalled();
+  });
+
+  it("flushes at sentence boundary once buffer >= threshold", async () => {
+    const onFlush = vi.fn().mockResolvedValue(undefined);
+    const buf = new SegmentedStreamBuffer(onFlush, 20);
+    await buf.append("Hello world. This is more text without boundary");
+    expect(onFlush).toHaveBeenCalledTimes(1);
+    expect(onFlush.mock.calls[0][0]).toBe("Hello world. ");
+    expect(buf.getBuffer()).toBe("This is more text without boundary");
+  });
+
+  it("flushRemaining drains the buffer", async () => {
+    const onFlush = vi.fn().mockResolvedValue(undefined);
+    const buf = new SegmentedStreamBuffer(onFlush, 1000);
+    await buf.append("leftover");
+    expect(onFlush).not.toHaveBeenCalled();
+    await buf.flushRemaining();
+    expect(onFlush).toHaveBeenCalledWith("leftover");
+    expect(buf.getBuffer()).toBe("");
+  });
+
+  it("flushRemaining is a no-op when buffer is empty", async () => {
+    const onFlush = vi.fn().mockResolvedValue(undefined);
+    const buf = new SegmentedStreamBuffer(onFlush, 100);
+    await buf.flushRemaining();
+    expect(onFlush).not.toHaveBeenCalled();
+  });
+});
+
+describe("chunkDiscordMessage", () => {
+  it("returns single chunk for short content", () => {
+    expect(chunkDiscordMessage("hi")).toEqual(["hi"]);
+  });
+
+  it("splits content larger than max length", () => {
+    const long = "x".repeat(2500);
+    const chunks = chunkDiscordMessage(long);
+    expect(chunks.length).toBeGreaterThan(1);
+    for (const c of chunks) {
+      expect(c.length).toBeLessThanOrEqual(2000);
+    }
+    expect(chunks.join("")).toBe(long);
+  });
+
+  it("prefers newline break points", () => {
+    const part = "a".repeat(1500) + "\n" + "b".repeat(1000);
+    const chunks = chunkDiscordMessage(part);
+    expect(chunks[0]).toBe("a".repeat(1500));
+    expect(chunks[1]).toBe("b".repeat(1000));
+  });
+});
 
 type SendMock = ReturnType<typeof vi.fn>;
 type ReplyMock = ReturnType<typeof vi.fn>;
